@@ -140,10 +140,14 @@ describe("The 'set' method", function () {
                         });
                     });
 
-                    // TODO: This test has to be refactored from v.3.0 on to expect an error instead
-                    it("converts null to 'null'", function (done) {
-                        client.set('foo', null);
-                        client.get('foo', helper.isString('null', done));
+                    it('errors if null value is passed', function (done) {
+                        try {
+                            client.set('foo', null);
+                            assert(false);
+                        } catch (error) {
+                            assert(/The SET command contains a invalid argument type./.test(error.message));
+                        }
+                        client.get('foo', helper.isNull(done));
                     });
 
                     it('emit an error with only the key set', function (done) {
